requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.8.1
    # via django
crispy-bootstrap5==2024.10
    # via -r requirements.in
django==5.1.4
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-crispy-forms
    #   django-extensions
    #   django-phonenumber-field
django-crispy-forms==2.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-extensions==3.2.3
    # via -r requirements.in
django-phonenumber-field==8.0.0
    # via -r requirements.in
phonenumberslite==8.13.52
    # via -r requirements.in
pillow==11.1.0
    # via -r requirements.in
sqlparse==0.5.3
    # via django
